Renamed Sieve to Shower.
The name was incorrect because a sieve *removes* incorrect elements, while the purpose of the shower is to *produce* elements that are likely correct.
A shower is basically a sieve from the other side. A real shower produces a few streams a high pressure, like a prime shower produces few elements that are more likely to be suitable.